requirements.txt
argparse
omegaconf
wandb
numpy
scipy
tqdm
torchrl
torchtune
tensordict
conda
ml-collections==0.1.1
torchao